Add autopkgtests
godot3-runner is built (by design) with tools=no, which means it doesn't
have any unit tests built-in. Between the editor and the server it
should be covered.

The editor test is marked as flaky as it seems to fail at different
tests across different runs on Salsa CI but passes fine locally.