Salsa CI: Intentionally disable BLHC and reprotest as unnecessary for now
Both of these were failing for various reasons. They are not important for
the general goals of packaging Godot right now, so instead of engineering
ways to get those jobs green, simply disable them so that overall status
is green, CI greenness becomes the expected norm, and any new regressions
will be caught.